# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Nagios Management Portal API",
  "version": "1.0.8"
 },
 "paths": {
...
  "/api/monitoring/hosts": {
   "get": {
    "tags": [
     "monitoring"
    ],
    "summary": "Mon Hosts",
    "operationId": "mon_hosts_api_monitoring_hosts_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 100
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "state",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 20
        },
        {
         "type": "null"
        }
       ],
       "title": "State"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/monitoring/services": {
   "get": {
    "tags": [
     "monitoring"
    ],
    "summary": "Mon Services",
    "operationId": "mon_services_api_monitoring_services_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 100
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "state",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 20
        },
        {
         "type": "null"
        }
       ],
       "title": "State"
      }
     },
     {
      "name": "host",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 64
        },
        {
         "type": "null"
        }
       ],
       "title": "Host"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 1000,
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/monitoring/problems": {
   "get": {
    "tags": [
     "monitoring"
    ],
    "summary": "Mon Problems",
    "operationId": "mon_problems_api_monitoring_problems_get",
    "parameters": [
     {
      "name": "include_handled",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Include Handled"
      }
     },
     {
      "name": "hide_on_down_hosts",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Hide On Down Hosts"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/monitoring/acknowledge": {
   "post": {
    "tags": [
     "monitoring"
    ],
    "summary": "Acknowledge",
    "operationId": "acknowledge_api_monitoring_acknowledge_post",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AckIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "monitoring"
    ],
    "summary": "Remove Ack",
    "operationId": "remove_ack_api_monitoring_acknowledge_delete",
    "parameters": [
     {
      "name": "host_name",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "maxLength": 64,
       "title": "Host Name"
      }
     },
     {
      "name": "service_description",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 100
        },
        {
         "type": "null"
        }
       ],
       "title": "Service Description"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/monitoring/downtime": {
   "post": {
    "tags": [
     "monitoring"
    ],
    "summary": "Schedule Downtime",
    "operationId": "schedule_downtime_api_monitoring_downtime_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DowntimeIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/monitoring/downtime/{kind}/{downtime_id}": {
   "delete": {
    "tags": [
     "monitoring"
    ],
    "summary": "Cancel Downtime",
    "operationId": "cancel_downtime_api_monitoring_downtime__kind___downtime_id__delete",
    "parameters": [
     {
      "name": "kind",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Kind"
      }
     },
     {
      "name": "downtime_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Downtime Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/monitoring/recheck": {
   "post": {
    "tags": [
     "monitoring"
    ],
    "summary": "Recheck",
    "operationId": "recheck_api_monitoring_recheck_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RecheckIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/events": {
   "get": {
    "tags": [
     "monitoring"
    ],
    "summary": "Events",
    "operationId": "events_api_events_get",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 2000,
       "minimum": 1,
       "default": 200,
       "title": "Limit"
      }
     },
     {
      "name": "host",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 64
        },
        {
         "type": "null"
        }
       ],
       "title": "Host"
      }
     },
     {
      "name": "kind",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 40
        },
        {
         "type": "null"
        }
       ],
       "title": "Kind"
      }
     },
     {
      "name": "hours",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer",
         "maximum": 744,
         "minimum": 1
        },
        {
         "type": "null"
        }
       ],
       "title": "Hours"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/dashboard": {
   "get": {
    "tags": [
     "monitoring"
    ],
    "summary": "Dashboard",
    "operationId": "dashboard_api_dashboard_get",
    "parameters": [
     {
      "name": "company_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Company Id"
      }
     },
     {
      "name": "location_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Location Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "AckIn": {
    "properties": {
     "host_name": {
      "type": "string",
      "maxLength": 64,
      "title": "Host Name"
     },
     "service_description": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 100
       },
       {
        "type": "null"
       }
      ],
      "title": "Service Description"
     },
     "comment": {
      "type": "string",
      "maxLength": 255,
      "minLength": 1,
      "title": "Comment"
     },
     "sticky": {
      "type": "boolean",
      "title": "Sticky",
      "default": true
     },
     "notify": {
      "type": "boolean",
      "title": "Notify",
      "default": true
     },
     "persistent": {
      "type": "boolean",
      "title": "Persistent",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "host_name",
     "comment"
    ],
    "title": "AckIn"
   },
...
   "DowntimeIn": {
    "properties": {
     "host_name": {
      "type": "string",
      "maxLength": 64,
      "title": "Host Name"
     },
     "service_description": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 100
       },
       {
        "type": "null"
       }
      ],
      "title": "Service Description"
     },
     "start": {
      "type": "string",
      "format": "date-time",
      "title": "Start"
     },
     "end": {
      "type": "string",
      "format": "date-time",
      "title": "End"
     },
     "comment": {
      "type": "string",
      "maxLength": 255,
      "minLength": 1,
      "title": "Comment"
     },
     "include_services": {
      "type": "boolean",
      "title": "Include Services",
      "default": true
     },
     "fixed": {
      "type": "boolean",
      "title": "Fixed",
      "default": true
     }
    },
    "type": "object",
    "required": [
     "host_name",
     "start",
     "end",
     "comment"
    ],
    "title": "DowntimeIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "RecheckIn": {
    "properties": {
     "host_name": {
      "type": "string",
      "maxLength": 64,
      "title": "Host Name"
     },
     "service_description": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 100
       },
       {
        "type": "null"
       }
      ],
      "title": "Service Description"
     },
     "all_services": {
      "type": "boolean",
      "title": "All Services",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "host_name"
    ],
    "title": "RecheckIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}